fix: adjust the ServiceParamAssignInput component
