+++ apps/web/src/pages/workflow/views/editor/components/config-panel/components/index.ts
@@ -16,4 +16,4 @@ export {
     type EntityFilterSelectValueType,
 } from './entity-filter-select';
 export { default as ParamAssignInput, type ParamAssignInputValueType } from './param-assign-input';
-export { default as ServiceEntityCall } from './service-entity-call';
+export { default as ServiceParamAssignInput } from './service-param-assign-input';
