--- apps/web/src/pages/workflow/views/editor/components/config-panel/components/service-entity-call/style.less
@@ -1,40 +1,0 @@
-.@{prefix}-service-entity-call {
-    &-title {
-        font-size: @font-size-lg;
-        font-weight: @font-weight-medium;
-        line-height: 24px;
-        color: var(--gray-9);
-    }
-
-    &-input {
-        display: block;
-
-        &-divider {
-            margin: @margin-xl 0 !important;
-        }
-
-        .param-item {
-            margin-top: @margin-md;
-
-            &-title {
-                margin-bottom: @margin-xs;
-            }
-
-            &-name {
-                font-size: @font-size-md;
-                font-weight: @font-weight-medium;
-                line-height: 22px;
-                color: #272e3b;
-            }
-
-            &-type {
-                margin-left: @margin-xs;
-                font-size: @font-size-md;
-                font-style: normal;
-                line-height: 22px;
-                color: #6b7785;
-            }
-        }
-    }
-
-}
