+++ apps/web/src/pages/workflow/views/editor/components/config-panel/components/service-param-assign-input/style.less
@@ -1,0 +1,24 @@
+.@{prefix}-service-param-assign-input {
+    .param-item {
+        margin-top: @margin-md;
+
+        &-title {
+            margin-bottom: @margin-xs;
+        }
+
+        &-name {
+            font-size: @font-size-md;
+            font-weight: @font-weight-medium;
+            line-height: 22px;
+            color: #272e3b;
+        }
+
+        &-type {
+            margin-left: @margin-xs;
+            font-size: @font-size-md;
+            font-style: normal;
+            line-height: 22px;
+            color: #6b7785;
+        }
+    }
+}
