+++ apps/web/src/pages/workflow/views/editor/components/config-panel/hooks/useNodeFormItems.tsx
@@ -1,4 +1,4 @@
-import { useMemo } from 'react';
+import { useMemo, useState } from 'react';
 import { type ControllerProps } from 'react-hook-form';
 import { TextField } from '@mui/material';
 import { useI18n } from '@milesight/shared/src/hooks';
@@ -14,7 +14,7 @@ import {
     // ParamInputSelect,
     TimerInput,
     ParamInput,
-    ServiceEntityCall,
+    ServiceParamAssignInput,
 } from '../components';
 
 type NodeFormGroupType = {
@@ -30,7 +30,7 @@ export type NodeFormDataProps = Record<string, any>;
 
 const useNodeFormItems = (node?: WorkflowNode) => {
     // const { getIntlText } = useI18n();
-
+    const [serviceKey, setServiceKey] = useState<ApiKey>();
     const formConfigs = useMemo(() => {
         const result: Partial<Record<WorkflowNodeType, NodeFormGroupType[]>> = {
             trigger: [
@@ -131,7 +131,10 @@ const useNodeFormItems = (node?: WorkflowNode) => {
                                     <EntitySelect
                                         filterModel={{ type: 'SERVICE' }}
                                         value={value}
-                                        onChange={onChange}
+                                        onChange={value => {
+                                            setServiceKey(value);
+                                            onChange(value);
+                                        }}
                                     />
                                 );
                             },
@@ -141,7 +144,20 @@ const useNodeFormItems = (node?: WorkflowNode) => {
                 {
                     groupName: 'Input Variables',
                     helperText: 'Please select the service you want to call first.',
-                    children: [],
+                    children: [
+                        {
+                            name: 'paramList',
+                            render({ field: { onChange, value } }) {
+                                return (
+                                    <ServiceParamAssignInput
+                                        serviceKey={serviceKey}
+                                        value={value}
+                                        onChange={onChange}
+                                    />
+                                );
+                            },
+                        },
+                    ],
                 },
             ],
             assigner: [
@@ -276,18 +292,10 @@ const useNodeFormItems = (node?: WorkflowNode) => {
                     ],
                 },
             ],
-            service: [
-                {
-                    name: 'service',
-                    render({ field: { onChange, value }, fieldState: { error } }) {
-                        return <ServiceEntityCall required value={value} onChange={onChange} />;
-                    },
-                },
-            ],
         };
 
         return result;
-    }, []);
+    }, [serviceKey]);
 
     return !node?.type ? [] : formConfigs[node.type] || [];
 };
